Added preprocessing script to maintain source/fortran structure while preprocessing. Added ability for such preprocessing to resolve dp issue

+++ source/fortran/global_variables.f90
@@ -54,7 +54,7 @@ module global_variables
   integer :: iscan_global
   !! Makes iscan available globally.
 
-  real(8) :: convergence_parameter
+  real(dp) :: convergence_parameter
   !! VMCON convergence parameter "sum"
 
   contains
